renamed into shared.rs
