+++ contracts/coordinator/src/lib.rs
@@ -4,4 +4,4 @@ pub use client::Client;
 pub mod contract;
 pub mod msg;
 mod state;
-pub mod exported;
+pub mod shared;
