+++ contracts/coordinator/src/shared.rs
